+++ test/spec/verbose.js
@@ -23,7 +23,11 @@ describe('verbose option', function() {
     });
 
     it('set data-bv-verbose="false" for form', function() {
-        var bv        = $('#verboseForm').attr('data-bv-verbose', 'false').bootstrapValidator().data('bootstrapValidator'),
+        var bv        = $('#verboseForm')
+                            .attr('data-bv-verbose', 'false')
+                            .bootstrapValidator('destroy')
+                            .bootstrapValidator()
+                            .data('bootstrapValidator'),
             $fullName = bv.getFieldElements('fullName'),
             messages;
 
@@ -34,7 +38,7 @@ describe('verbose option', function() {
         expect(messages[0]).toEqual($fullName.attr('data-bv-notempty-message'));
 
         bv.resetForm();
-        $fullName.val('Special@#$');
+        $fullName.val('Spe@#$');
         bv.validate();
         messages = bv.getMessages('fullName');
         expect(messages.length).toEqual(1);
@@ -54,7 +58,9 @@ describe('verbose option', function() {
                             .find('[name="fullName"]')
                                 .attr('data-bv-verbose', 'false')
                                 .end()
-                            .bootstrapValidator().data('bootstrapValidator'),
+                            .bootstrapValidator('destroy')
+                            .bootstrapValidator()
+                            .data('bootstrapValidator'),
             $fullName = bv.getFieldElements('fullName'),
             messages;
 
@@ -65,7 +71,7 @@ describe('verbose option', function() {
         expect(messages[0]).toEqual($fullName.attr('data-bv-notempty-message'));
 
         bv.resetForm();
-        $fullName.val('Special@#$');
+        $fullName.val('Spe@#$');
         bv.validate();
         messages = bv.getMessages('fullName');
         expect(messages.length).toEqual(1);
@@ -80,7 +86,10 @@ describe('verbose option', function() {
     });
 
     it('set verbose: "false" for form', function() {
-        var bv        = $('#verboseForm').bootstrapValidator({ verbose: false }).data('bootstrapValidator'),
+        var bv        = $('#verboseForm')
+                            .bootstrapValidator('destroy')
+                            .bootstrapValidator({ verbose: false })
+                            .data('bootstrapValidator'),
             $fullName = bv.getFieldElements('fullName'),
             messages;
 
@@ -91,7 +100,7 @@ describe('verbose option', function() {
         expect(messages[0]).toEqual($fullName.attr('data-bv-notempty-message'));
 
         bv.resetForm();
-        $fullName.val('Special@#$');
+        $fullName.val('Spe@#$');
         bv.validate();
         messages = bv.getMessages('fullName');
         expect(messages.length).toEqual(1);
@@ -105,9 +114,11 @@ describe('verbose option', function() {
         expect(messages[0]).toEqual($fullName.attr('data-bv-stringlength-message'));
     });
 
+    // #1057
     it('set verbose: "false" for field', function() {
         var bv        = $('#verboseForm')
                             .attr('data-bv-verbose', 'true')
+                            .bootstrapValidator('destroy')
                             .bootstrapValidator({
                                 verbose: true,
                                 fields: {
@@ -127,7 +138,7 @@ describe('verbose option', function() {
         expect(messages[0]).toEqual($fullName.attr('data-bv-notempty-message'));
 
         bv.resetForm();
-        $fullName.val('Special@#$');
+        $fullName.val('Spe@#$');
         bv.validate();
         messages = bv.getMessages('fullName');
         expect(messages.length).toEqual(1);
@@ -139,5 +150,45 @@ describe('verbose option', function() {
         messages = bv.getMessages('fullName');
         expect(messages.length).toEqual(1);
         expect(messages[0]).toEqual($fullName.attr('data-bv-stringlength-message'));
+    });
+
+    // #1055
+    it('trigger "error.field.bv" event', function() {
+        var validators = [],    // Array of not passed validators
+            bv         = $('#verboseForm')
+                            .attr('data-bv-verbose', 'true')
+                            .bootstrapValidator('destroy')
+                            .bootstrapValidator({
+                                verbose: true,
+                                fields: {
+                                    fullName: {
+                                        verbose: false
+                                    }
+                                }
+                            })
+                            .on('error.field.bv', function(e, data) {
+                                validators.push(data.validator);
+                            })
+                            .data('bootstrapValidator'),
+            $fullName  = bv.getFieldElements('fullName');
+
+        $fullName.val('');
+        bv.validate();
+        expect(validators.length).toEqual(1);
+        expect(validators[0]).toEqual('notEmpty');
+
+        validators = [];
+        bv.resetForm();
+        $fullName.val('Spe@#$');
+        bv.validate();
+        expect(validators.length).toEqual(1);
+        expect(validators[0]).toEqual('regexp');
+
+        validators = [];
+        bv.resetForm();
+        $fullName.val('Full');
+        bv.validate();
+        expect(validators.length).toEqual(1);
+        expect(validators[0]).toEqual('stringLength');
     });
 });